refactor: move ui modules
