--- apps/meteor/app/ui/client/lib/KonchatNotification.ts
@@ -1,1 +1,0 @@
-// KonchatNotification in memoriam
