+++ apps/meteor/client/lib/videoRecorder.ts
