+++ apps/meteor/client/views/room/providers/RoomProvider.tsx
@@ -7,12 +7,12 @@ import RoomToolboxProvider from './RoomToolboxProvider';
 import UserCardProvider from './UserCardProvider';
 import { useRedirectOnSettingsChanged } from './hooks/useRedirectOnSettingsChanged';
 import { useUsersNameChanged } from './hooks/useUsersNameChanged';
-import { UserAction } from '../../../../app/ui/client/lib/UserAction';
 import { omit } from '../../../../lib/utils/omit';
 import { useFireGlobalEvent } from '../../../hooks/useFireGlobalEvent';
 import { useRoomRolesQuery } from '../../../hooks/useRoomRolesQuery';
 import { RoomHistoryManager, useRoomHistoryState } from '../../../lib/RoomHistoryManager';
 import { RoomManager } from '../../../lib/RoomManager';
+import { UserAction } from '../../../lib/UserAction';
 import { roomCoordinator } from '../../../lib/rooms/roomCoordinator';
 import ImageGalleryProvider from '../../../providers/ImageGalleryProvider';
 import { Rooms, Subscriptions } from '../../../stores';
