+++ apps/meteor/client/views/room/providers/hooks/useChatMessagesInstance.spec.ts
@@ -4,8 +4,8 @@ import { useUserId } from '@rocket.chat/ui-contexts';
 import { renderHook } from '@testing-library/react';
 
 import { useChatMessagesInstance } from './useChatMessagesInstance';
-import { ChatMessages } from '../../../../../app/ui/client/lib/ChatMessages';
 import { useEmojiPicker } from '../../../../contexts/EmojiPickerContext';
+import { ChatMessages } from '../../../../lib/ChatMessages';
 import type { E2ERoomState } from '../../../../lib/e2ee/E2ERoomState';
 import { useUiKitActionManager } from '../../../../uikit/hooks/useUiKitActionManager';
 import { useRoomSubscription } from '../../contexts/RoomContext';
@@ -28,7 +28,7 @@ jest.mock('../../../../contexts/EmojiPickerContext', () => ({
 }));
 
 const updateSubscriptionMock = jest.fn();
-jest.mock('../../../../../app/ui/client/lib/ChatMessages', () => {
+jest.mock('../../../../lib/ChatMessages', () => {
 	return {
 		ChatMessages: jest.fn().mockImplementation(() => {
 			return {
