+++ apps/meteor/client/views/room/providers/hooks/useChatMessagesInstance.ts
@@ -3,8 +3,8 @@ import { useUserId } from '@rocket.chat/ui-contexts';
 import { useEffect } from 'react';
 
 import { useInstance } from './useInstance';
-import { ChatMessages } from '../../../../../app/ui/client/lib/ChatMessages';
 import { useEmojiPicker } from '../../../../contexts/EmojiPickerContext';
+import { ChatMessages } from '../../../../lib/ChatMessages';
 import type { ChatAPI } from '../../../../lib/chats/ChatAPI';
 import { useUiKitActionManager } from '../../../../uikit/hooks/useUiKitActionManager';
 import { useRoomSubscription } from '../../contexts/RoomContext';
